requirement.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
chardet==3.0.4
discord.py==1.6.0
idna==3.1
lxml==4.6.2
MouseInfo==0.1.3
multidict==5.1.0
Pillow==8.1.0
pkg-resources==0.0.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-dotenv==0.15.0
python3-xlib==0.15
PyTweening==1.0.3
selenium==3.141.0
soupsieve==2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
